fix: 코드 수정 및 주석 제거 및 파일 이동
