+++ src/main/java/com/avengers/musinsa/viewController/CartViewController.java
@@ -1,4 +1,4 @@
-package com.avengers.musinsa;
+package com.avengers.musinsa.viewController;
 
 import com.avengers.musinsa.domain.user.auth.jwt.TokenProviderService;
 import com.avengers.musinsa.domain.user.dto.ProductsInCartInfoResponse;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 @RequiredArgsConstructor
-public class ViewController {
+public class CartViewController {
 
     private final CartService cartService;
