+++ src/main/java/com/avengers/musinsa/viewController/ProductViewController.java
@@ -1,0 +1,50 @@
+package com.avengers.musinsa.viewController;
+
+import com.avengers.musinsa.domain.product.dto.search.SearchResponse;
+import com.avengers.musinsa.domain.product.service.ProductService;
+import com.avengers.musinsa.domain.product.service.ProductServiceImpl;
+import com.avengers.musinsa.domain.user.auth.jwt.TokenProviderService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+@RequestMapping("/products")
+@RequiredArgsConstructor
+public class ProductViewController {
+    private final TokenProviderService tokenProviderService;
+    private final ProductService productService;
+
+    @GetMapping("/{productId}")
+    public String geetProductDetail(@PathVariable Long productId, Model model) {
+
+        return "product/productDetailPage";
+    }
+
+    // /search?keyword=...  또는 /search/?keyword=...
+    @GetMapping()
+    public String searchPage(
+            @RequestParam(name = "keyword", required = false, defaultValue = "") String keyword,
+            @CookieValue(value = "Authorization", required = false) String authorization,
+            Model model
+    ) {
+        Long userId = null;
+        if (authorization != null && !authorization.isBlank()) {
+            try {
+                userId = tokenProviderService.getUserIdFromToken(authorization);
+            } catch (Exception ignore) { /* 비로그인/만료 등은 null 로 */ }
+        }
+
+        SearchResponse result = productService.searchProducts(keyword, userId);
+
+        model.addAttribute("keyword", keyword);
+        model.addAttribute("result", result);
+        // JSP에서 ${result} 사용
+        return "product/searchProducts"; // JSP 뷰 경로
+    }
+}
